docker-tensorfood/requirements.txt
scikit-learn==0.23.1
numpy==1.18.5    
Flask==1.1.2  
waitress==1.4.4
pandas==1.0.5
Pillow==7.2.0
tensorflow==2.3.0